スクリーンショットアプリ/requirements.txt
pygetwindow
pywin32
pillow
screeninfo
pywinauto
